moved `EvalErr` to own file
