+++ src/lurk/mod.rs
@@ -3,6 +3,7 @@
 pub mod big_num;
 pub mod chipset;
 pub mod cli;
+pub mod error;
 pub mod eval;
 pub mod ingress;
 pub mod lang;
